feat(ui): merge main_window into main
